Update upgrade.php
The problem is that when building the SQL at the end of a line put a carriage return and continued on another line by the end of the SQL statement. 

Regardless of the database (mysql or postgres or oracle ...) is used, it is best to put that construction of SQL statements on a single line, without adding a carriage return.